WIP (ng-content without new component) - duplacated code
