+++ apps/catalog/catalog/src/app/dashboard/title/view/view.component.html
@@ -5,6 +5,50 @@
   <title-dashboard-shell [routes]="navLinks">
     <movie-cta>
       <!-- TODO #6084 CTA -->
+      <aside fxLayout="column" fxLayout.lt-md="row" fxLayoutAlign="space-between end" fxFlex>
+        <ng-container *ngIf="movie.app[appName].status === 'archived' else cta">
+          <button mat-button color="primary" (click)="updateStatus('draft', 'Title back to Draft')">
+            <mat-icon svgIcon="restore"></mat-icon>
+            Restore Title
+          </button>
+        </ng-container>
+        <ng-template #cta>
+          <ng-container *ngIf="org$ | async | canAccess: 'marketplace': appName">
+            <a mat-button color="primary" [routerLink]="['/c/o/marketplace/title', movie.id, 'main']">
+              <mat-icon svgIcon="visibility" color="primary"></mat-icon>
+              <span>Marketplace Preview</span>
+            </a>
+          </ng-container>
+        </ng-template>
+        <!-- TODO #6084 CTA -->
+        <ng-container *ngIf="movie.app[appName].status === 'archived' else actionsMenu">
+          <button mat-button (click)="removeAppAccess()" matTooltip="Delete Title">
+            <mat-icon svgIcon="delete"></mat-icon>
+            <span>Delete Title</span>
+          </button>
+        </ng-container>
+        <ng-template #actionsMenu>
+          <button mat-button [matMenuTriggerFor]="actionMenuMenu">
+            <mat-icon svgIcon="more_horiz"></mat-icon>
+            <span>Actions</span>
+            <mat-icon svgIcon="arrow_drop_down"></mat-icon>
+          </button>
+          <mat-menu #actionMenuMenu="matMenu">
+            <a mat-menu-item [routerLink]="['/c/o/dashboard/tunnel/movie', movie.id]">
+              <mat-icon svgIcon="edit"></mat-icon><span>Edit Title</span>
+            </a>
+            <button *ngIf="movie.app[appName].status !== 'draft'" mat-menu-item (click)="updateStatus('draft', 'Title back to Draft.')">
+              <mat-icon svgIcon="drafts"></mat-icon><span>Switch back to Draft</span>
+            </button>
+            <button mat-menu-item (click)="updateStatus('archived', 'Title archived.')">
+              <mat-icon svgIcon="archive"></mat-icon><span>Archive Title</span>
+            </button>
+            <a mat-menu-item *ngIf="['catalog','financiers'].includes(appName) && movie.app[appName].status === 'draft'"  [routerLink]="['/c/o/dashboard/tunnel/movie', movie.id, 'summary']"  fragment="submit">
+              <mat-icon svgIcon="cloud_upload"></mat-icon><span>Submit Title</span>
+            </a>
+          </mat-menu>
+        </ng-template>
+      </aside>
     </movie-cta>
   </title-dashboard-shell>
 </ng-container>
