+++ apps/catalog/catalog/src/app/dashboard/title/view/view.component.ts
@@ -6,6 +6,14 @@ import { Router, NavigationEnd } from '@angular/router';
 import { Component, OnInit, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { RouteDescription } from '@blockframes/utils/common-interfaces/navigation';
+import { RouterQuery } from '@datorama/akita-ng-router-store';
+import { MatDialog } from '@angular/material/dialog';
+import { MovieService } from '@blockframes/movie/+state';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { OrganizationQuery } from '@blockframes/organization/+state';
+import { getAppName, getCurrentApp, getMovieAppAccess } from '@blockframes/utils/apps';
+import { ConfirmInputComponent } from '@blockframes/ui/confirm-input/confirm-input.component';
+import { storeStatus, StoreStatus } from '@blockframes/utils/static-model';
 
 @Component({
   selector: 'catalog-title-view',
@@ -17,6 +25,9 @@ export class TitleViewComponent implements OnInit, OnDestroy {
   public movie$: Observable<Movie>;
   public loading$: Observable<boolean>;
   private sub: Subscription;
+
+  public org$ = this.orgQuery.selectActive();
+  public appName = getCurrentApp(this.routerQuery);
 
   navLinks: RouteDescription[] = [
     {
@@ -41,7 +52,17 @@ export class TitleViewComponent implements OnInit, OnDestroy {
     }
   ];
 
-  constructor(private movieQuery: MovieQuery, private dynTitle: DynamicTitleService, private router: Router) {
+  constructor(
+    private movieQuery: MovieQuery,
+    private dynTitle: DynamicTitleService,
+    private routerQuery: RouterQuery,
+    private query: MovieQuery,
+    private dialog: MatDialog,
+    private movieService: MovieService,
+    private snackBar: MatSnackBar,
+    private router: Router,
+    private orgQuery: OrganizationQuery
+    ) {
     const titleName = this.movieQuery.getActive().title.international || 'No title'
     this.sub = this.router.events.subscribe(event => {
       if (event instanceof NavigationEnd) {
@@ -63,5 +84,54 @@ export class TitleViewComponent implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.sub.unsubscribe();
+  }
+
+  removeAppAccess() {
+    const movie = this.query.getActive();
+    const appsName = getMovieAppAccess(movie).map(a => getAppName(a).label);
+    this.dialog.open(ConfirmInputComponent, {
+      data: {
+        title: `You are about to delete ${movie.title.international} permanently.`,
+        subtitle: `This Title will still be available on <i>${appsName.join(', ')}</i>.<br/> If you wish to proceed, please type "DELETE" in the field below.`,
+        confirmationWord: 'delete',
+        confirmButtonText: 'delete title',
+        cancelButtonText: 'keep title',
+        onConfirm: async () => {
+          await this.movieService.update(movie.id, movie => ({
+            ...movie,
+            app: {
+              ...movie.app,
+              [this.appName]: {
+                ...movie.app[this.appName],
+                access: false
+              }
+            }
+          }));
+
+          const ref = this.snackBar.open('Title deleted.', '', { duration: 4000 });
+          ref.afterDismissed().subscribe(() => this.router.navigate(['/c/o/dashboard/title']));
+        }
+      }
+    })
+  }
+
+  async updateStatus(status: StoreStatus, message?: string) {
+    const movie = this.query.getActive();
+    await this.movieService.update(movie.id, movie => ({
+      ...movie,
+      app: {
+        ...movie.app,
+        [this.appName]: {
+          ...movie.app[this.appName],
+          status: status
+        }
+      }
+    }));
+
+    if (message) {
+      this.snackBar.open(message, '', { duration: 4000 });
+    } else {
+      this.snackBar.open(`Title ${storeStatus[status]}.`, '', { duration: 4000 });
+    }
   }
 }
