+++ apps/festival/festival/src/app/dashboard/title/view/view.component.ts
@@ -3,6 +3,15 @@ import { Observable } from 'rxjs';
 import { Movie } from '@blockframes/movie/+state/movie.model';
 import { MovieQuery } from '@blockframes/movie/+state/movie.query';
 import { RouteDescription } from '@blockframes/utils/common-interfaces/navigation';
+import { getAppName, getCurrentApp, getMovieAppAccess } from '@blockframes/utils/apps';
+import { RouterQuery } from '@datorama/akita-ng-router-store';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmInputComponent } from '@blockframes/ui/confirm-input/confirm-input.component';
+import { MovieService } from '@blockframes/movie/+state';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { storeStatus, StoreStatus } from '@blockframes/utils/static-model';
+import { Router } from '@angular/router';
+import { OrganizationQuery } from '@blockframes/organization/+state';
 
 @Component({
   selector: 'festival-dashboard-title-view',
@@ -14,6 +23,9 @@ import { RouteDescription } from '@blockframes/utils/common-interfaces/navigatio
 export class TitleViewComponent implements OnInit {
   public movie$: Observable<Movie>;
   public loading$: Observable<boolean>;
+
+  public org$ = this.orgQuery.selectActive();
+  public appName = getCurrentApp(this.routerQuery);
 
   navLinks: RouteDescription[] = [
     {
@@ -38,7 +50,16 @@ export class TitleViewComponent implements OnInit {
     }
   ];
 
-  constructor(private movieQuery: MovieQuery) {}
+  constructor(
+    private movieQuery: MovieQuery,
+    private routerQuery: RouterQuery,
+    private query: MovieQuery,
+    private dialog: MatDialog,
+    private movieService: MovieService,
+    private snackBar: MatSnackBar,
+    private router: Router,
+    private orgQuery: OrganizationQuery
+    ) {}
 
   ngOnInit() {
     this.getMovie();
@@ -47,5 +68,55 @@ export class TitleViewComponent implements OnInit {
   private getMovie() {
     this.loading$ = this.movieQuery.selectLoading();
     this.movie$ = this.movieQuery.selectActive();
+  }
+
+
+  removeAppAccess() {
+    const movie = this.query.getActive();
+    const appsName = getMovieAppAccess(movie).map(a => getAppName(a).label);
+    this.dialog.open(ConfirmInputComponent, {
+      data: {
+        title: `You are about to delete ${movie.title.international} permanently.`,
+        subtitle: `This Title will still be available on <i>${appsName.join(', ')}</i>.<br/> If you wish to proceed, please type "DELETE" in the field below.`,
+        confirmationWord: 'delete',
+        confirmButtonText: 'delete title',
+        cancelButtonText: 'keep title',
+        onConfirm: async () => {
+          await this.movieService.update(movie.id, movie => ({
+            ...movie,
+            app: {
+              ...movie.app,
+              [this.appName]: {
+                ...movie.app[this.appName],
+                access: false
+              }
+            }
+          }));
+
+          const ref = this.snackBar.open('Title deleted.', '', { duration: 4000 });
+          ref.afterDismissed().subscribe(() => this.router.navigate(['/c/o/dashboard/title']));
+        }
+      }
+    })
+  }
+
+  async updateStatus(status: StoreStatus, message?: string) {
+    const movie = this.query.getActive();
+    await this.movieService.update(movie.id, movie => ({
+      ...movie,
+      app: {
+        ...movie.app,
+        [this.appName]: {
+          ...movie.app[this.appName],
+          status: status
+        }
+      }
+    }));
+
+    if (message) {
+      this.snackBar.open(message, '', { duration: 4000 });
+    } else {
+      this.snackBar.open(`Title ${storeStatus[status]}.`, '', { duration: 4000 });
+    }
   }
 }
