+++ apps/financiers/financiers/src/app/dashboard/title/view/view.component.scss
@@ -2,3 +2,9 @@ header {
   text-align: center;
   margin-bottom: 24px;
 }
+
+movie-cta {
+  span {
+    // align-self: flex-start;
+  }
+}
