+++ apps/financiers/financiers/src/app/dashboard/title/view/view.component.ts
@@ -9,6 +9,13 @@ import { DashboardTitleShellComponent } from '@blockframes/movie/dashboard/shell
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { CampaignService } from '@blockframes/campaign/+state';
 import { filter, switchMap } from 'rxjs/operators';
+import { RouterQuery } from '@datorama/akita-ng-router-store';
+import { Router } from '@angular/router';
+import { OrganizationQuery } from '@blockframes/organization/+state';
+import { MovieService } from '@blockframes/movie/+state';
+import { getAppName, getCurrentApp, getMovieAppAccess } from '@blockframes/utils/apps';
+import { ConfirmInputComponent } from '@blockframes/ui/confirm-input/confirm-input.component';
+import { storeStatus, StoreStatus } from '@blockframes/utils/static-model';
 
 
 const links: RouteDescription[] = [
@@ -58,11 +65,18 @@ export class TitleViewComponent implements OnInit, OnDestroy {
   public sub: Subscription;
   public percentage = 0;
 
+  public org$ = this.orgQuery.selectActive();
+  public appName = getCurrentApp(this.routerQuery);
+
   constructor(
     private movieQuery: MovieQuery,
     private dialog: MatDialog,
+    private routerQuery: RouterQuery,
     private snackbar: MatSnackBar,
     private campaignService: CampaignService,
+    private movieService: MovieService,
+    private router: Router,
+    private orgQuery: OrganizationQuery,
     private cdr: ChangeDetectorRef
   ) { }
 
@@ -95,5 +109,54 @@ export class TitleViewComponent implements OnInit, OnDestroy {
     this.dialogRef.close();
     await this.shell.getConfig('campaign').onSave();
     this.snackbar.open('The funding status has been updated.', null, { duration: 1000 });
+  }
+
+  removeAppAccess() {
+    const movie = this.movieQuery.getActive();
+    const appsName = getMovieAppAccess(movie).map(a => getAppName(a).label);
+    this.dialog.open(ConfirmInputComponent, {
+      data: {
+        title: `You are about to delete ${movie.title.international} permanently.`,
+        subtitle: `This Title will still be available on <i>${appsName.join(', ')}</i>.<br/> If you wish to proceed, please type "DELETE" in the field below.`,
+        confirmationWord: 'delete',
+        confirmButtonText: 'delete title',
+        cancelButtonText: 'keep title',
+        onConfirm: async () => {
+          await this.movieService.update(movie.id, movie => ({
+            ...movie,
+            app: {
+              ...movie.app,
+              [this.appName]: {
+                ...movie.app[this.appName],
+                access: false
+              }
+            }
+          }));
+
+          const ref = this.snackbar.open('Title deleted.', '', { duration: 4000 });
+          ref.afterDismissed().subscribe(() => this.router.navigate(['/c/o/dashboard/title']));
+        }
+      }
+    })
+  }
+
+  async updateStatus(status: StoreStatus, message?: string) {
+    const movie = this.movieQuery.getActive();
+    await this.movieService.update(movie.id, movie => ({
+      ...movie,
+      app: {
+        ...movie.app,
+        [this.appName]: {
+          ...movie.app[this.appName],
+          status: status
+        }
+      }
+    }));
+
+    if (message) {
+      this.snackbar.open(message, '', { duration: 4000 });
+    } else {
+      this.snackbar.open(`Title ${storeStatus[status]}.`, '', { duration: 4000 });
+    }
   }
 }
